refine

+++ docker/ci/test/build.sh
@@ -10,5 +10,8 @@ proxy_args+=" --build-arg HTTPS_PROXY=${HTTPS_PROXY}"
 proxy_args+=" --build-arg http_proxy=${http_proxy}"
 proxy_args+=" --build-arg https_proxy=${https_proxy}"
 
-docker build --rm $proxy_args \
-    -t oneflow-test:$USER .
+img_tag="oneflow-test:0.1"
+if [[ "$(docker images -q ${img_tag} 2> /dev/null)" == "" ]]; then
+  docker build --rm $proxy_args \
+    -t $img_tag .
+fi
